streamlit/requirements.txt
streamlit==1.41.1
pandas==2.2.3
wordcloud~=1.9.4
requests~=2.32.3
numpy~=2.2.1
plotly~=5.24.1
nltk~=3.9.1
pillow~=11.0.0